mychatbot/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
argcomplete==3.1.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.3
attrs==23.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudevents==1.9.0
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
dataclasses-json==0.5.14
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
deprecation==2.1.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
faiss-cpu==1.7.4
Flask==2.3.3
frozenlist==1.4.0
functions-framework==3.4.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.34
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
ipython @ file:///C:/b/abs_254uk73z5b/croot/ipython_1691532131313/work
itsdangerous==2.1.2
jedi @ file:///C:/ci/jedi_1644315428289/work
Jinja2==3.1.2
langchain==0.0.279
langsmith==0.0.33
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661915841596/work
mime==0.1.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.10
packaging==23.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pep517==0.13.0
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pipx==1.2.0
prompt-toolkit @ file:///C:/b/abs_6coz5_9f2s/croot/prompt-toolkit_1672387908312/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pydantic==2.3.0
pydantic_core==2.6.3
Pygments @ file:///C:/b/abs_fay9dpq4n_/croot/pygments_1684279990574/work
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
SQLAlchemy==2.0.20
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
textbase-client==0.1.3
tiktoken==0.4.0
tomli==2.0.1
tqdm==4.66.1
traitlets @ file:///C:/b/abs_e5m_xjjl94/croot/traitlets_1671143896266/work
typing-inspect==0.9.0
typing_extensions @ file:///C:/b/abs_213vg2cd1l/croot/typing_extensions_1690297804941/work
urllib3==2.0.4
userpath==1.9.0
watchdog==3.0.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==2.3.7
yarl==1.9.2
yaspin==3.0.0
zipp==3.16.2
